Use Matchdep module to load installed Grunt plugins
Calling `loadNpmTasks` for each Grunt plugin installed is duplicative, so
Tyler Kellen created a matchdep package to allow Grunt to piggyback on
the list of Grunt plugins saved in package.json. Using this module reduces
boilerplate and makes it more obvious the package.json is the canonical
list of installed plugins. Tyler's a Grunt team member and Ben's endorsed
the syntax, so it's an officially-sanctioned approach.